+++ cs3560cli/main.py
@@ -1,13 +1,15 @@
 import click
 
-from .functions.check_username import check_username
-from .functions.watch_zip import watch_zip
-from .functions.create_gitignore import create_gitignore
-from .functions.highlight import highlight
+from .commands.check_username import check_username
+from .commands.create_gitignore import create_gitignore
+from .commands.highlight import highlight
+from .commands.watch_zip import watch_zip
+
 
 @click.group()
 def cli():
     pass
+
 
 cli.add_command(check_username)
 cli.add_command(watch_zip)